Fix swipe issue with EarnBadgeModal (#289)
When GoldBadgeInfoModal or GoldBadgeRewardModal displays on top of
EarnBadgeModal, the swipe event strangely bypasses the top modal and
fires on EarnBadgeModal. This change disables swipe gesture on the
top modal when two modals are stacked on top of each other.

However, this fix should be temporary. In the future, we should avoid
stacking two modals together.